Extract TEC-1 memory panel HTML
